Updated URLconf for books
